Added view controller for name entry
This controller gets name from user does a simple check to make sure the name isn't empty.
If it is empty there will be a error message telling the user to enter a name.
Once the user enters a name, it saves that name to UserDefaults and calls the didEnterName protocol method.
The keyboard is also dismissed, error message is set back to empty and first time setup is set to false